TicTacToeBoard

Documentation

convertFieldToString :: Field -> String

convertFieldToString takes a Field and returns its String representation on displayed board. For Cross it returns X, for Circle O end por Empty just space " ". (See Field)

boardToString :: Board -> String

boardToString takes the Board and returns its String representation. (See Board)

eliminate :: Maybe Board -> Board

eliminate converts Maybe Board into Board. If the Board is Nothing returns newBoard (See newBoard)

help :: String

help returns String in which whole board is displayed with numbers to help player choose the field they are going to fill.

eliminate2 :: Maybe Int -> Int

eliminate2 converts Maybe Int int Int