TicTacToeEngine

Documentation

type Board = Array (Int, Int) Field

Board represents the board for the Tic Tac Toe game. Starting board is represented by newBoard, which is an empty 3x3 Tic Tac Toe board. It's rows and columns are numbered from 0 to 2. Those numbers are used to specify where do you want to move. (See makeAMove for the exact way to specify where do you want to place your mark)

data Field

Field describes a possible state of a field on the playing board (represented by Board). A Field can be Empty, can have a Cross or a Circle.

Constructors

Empty

Cross

Circle

data GameStatus

GameStatus is a variable describing the state of the game - did it end or not and what is a result? InProgress indicates that the game hasn't ended yet. CircleWon, CrossWon an Stalemate indicate both the fact that the game has ended and the result.

Constructors

InProgress

CrossWon

CircleWon

Stalemate

newBoard :: Board

makeAMove :: Board -> Field -> (Int, Int) -> Maybe Board

makeAMove returns Just Board after making a specified move or Nothing if move is impossible. Move is specified by the current game represented by a Board, a Field indicating whether we want to place a Circle or a Cross and (Int,Int) index (i,j) with i representing a row number and j representing a column number. Example: makeAMove board Circle (0,2) "puts" a Circle in field in row 0, column 2, if possible, returns Nothing if impossible. (See Board for the layout of the board)

seeGameStatus :: Board -> GameStatus

takes a Board, returns information about status (GameStatus): is it won by circles, by crosses, is it a stalemate or is it ongoing? (See GameStatus)