TicTacToePlayer

Documentation

sensiblyDecideHowToMove :: Field -> Board -> Maybe Board

sensiblyDecideHowToMove takes a Board and a Field representing whether it is a Cross or a Circle move. It returns a new Board (either as a Just Board or Nothing if there is no possible move) after making a move that should either seize victory if it's immediately available, prevent the enemy from doing the same or - if none of those are available - simply places a required Cross or Circle in the first possible place.

lazilyDecideHowToMove :: Field -> Board -> Maybe Board

lazilyDecideHowToMove takes a Board and a Field representing whether it is a Cross or a Circle move. It returns a new Board after making a first possible move - either as a Just Board or Nothing if there is no possible move.

oneDimIndexToTwoDim :: Int -> (Int, Int)

oneDimIndexToTwoDim converts number of field into two dimensional index.

otherPlayersField :: Field -> Field

otherPlayersField returns next type of sign for the next player. Next for Cross is Circle and vice versa.

firstEmptyPlace :: Board -> (Int, Int)

firstEmptyPlace searches for first Empty place in the Board